requirements.txt
alembic==1.17.1
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
fastapi==0.119.0
filelock==3.20.0
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
idna==3.11
jiter==0.11.1
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.2.6
openai==2.6.1
opencv-python==4.12.0.88
packaging==25.0
pgvector==0.4.1
pillow==12.0.0
psycopg2==2.9.11
pydantic==2.12.1
pydantic_core==2.41.3
pytesseract==0.3.13
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.48.0
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
